Move cachedEventsMap to its own file
